Add methods to ImageService. Add PaginatedResult and SearchCriteria for all entities. Add DTOs for ImageResponse and ImageThumbResponse. Edit repositories.

+++ ImageCollection/src/main/java/pl/polsl/dsa/imagecollection/dao/CategoryRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import pl.polsl.dsa.imagecollection.model.CategoryEntity;
 
 public interface CategoryRepository extends
-        CrudRepository<CategoryEntity, Integer>,
+        CrudRepository<CategoryEntity, Long>,
         JpaSpecificationExecutor<CategoryEntity> {
 
 }
